requirements.txt
discord
requests
netml
argcmdr
argparse_formatter
terminaltables
PyYAML
joblib
xgboost
catboost
scapy
#python-dateutil
#pytz
missingno
imblearn
pybind11
python-dotenv
paramiko
flask
pytest
pyshark
elasticsearch
geoip2fast
flagpy
Flask-Login 
Flask-SQLAlchemy
Werkzeug
Flask-WTF
elasticsearch-dsl[async]
pycountry
PyJWT
dash
setuptools
dash_bootstrap_components
dash_ag_grid
geoip2
